Rename : MyPage 폴더 분리
- 마이페이지 관련 컴포넌트들 폴더를 만들어서 분리했습니다
