+++ src/components/MyPage/ManageCard.jsx
@@ -3,10 +3,6 @@ import React from 'react'
 const ManageCard = ({ userStatus }) => {
   const getTitle = () => {
     switch (userStatus) {
-      case 'leader':
-        return '결제 정산 관리'
-      case 'member':
-        return '결제 정산 관리'
       case 'none':
         return '결제 매칭 신청'
       default:
